Organize core files into lib/core/
