+++ src/lib/assets/pods.test.ts
@@ -1,7 +1,7 @@
 import { getNamespace, getWatcher, getDeployment, getModuleSecret, genEnv } from "./pods";
 import { expect, describe, test, jest, afterEach } from "@jest/globals";
 import { Assets } from ".";
-import { ModuleConfig } from "../module";
+import { ModuleConfig } from "../core/module";
 import { gzipSync } from "zlib";
 
 jest.mock("zlib");
