+++ src/lib/processors/watch-processor.test.ts
@@ -4,7 +4,7 @@ import { afterAll, beforeEach, describe, expect, it, jest } from "@jest/globals"
 import { GenericClass, K8s, KubernetesObject, kind } from "kubernetes-fluent-client";
 import { K8sInit, WatchPhase } from "kubernetes-fluent-client/dist/fluent/types";
 import { WatchCfg, WatchEvent, Watcher } from "kubernetes-fluent-client/dist/fluent/watch";
-import { Capability } from "../capability";
+import { Capability } from "../core/capability";
 import { setupWatch, logEvent, queueKey, getOrCreateQueue } from "./watch-processor";
 import Log from "../telemetry/logger";
 import { metricsCollector } from "../telemetry/metrics";
